0.5
Set up with the final changes needed to make the last of the stuff work in the draft of the rules. Finally got that stupid autosoft thing figured out. Turns out I'm a complete knobhead trumpet and was editing the version of the file in github desktop, not the one chummer actually *uses*.

And before you laugh at my sheer stupidity, I don't see you making amends here.