Make Reagent xpath less sloppy
Changed the xpath used to hide the reagents we don't want to be a series of xpaths that will specifically only target the reagents we want hidden. Because hiding all reagents and then overriding that with the append operation below that only works in STE and doesn't work in Point Buy. For some reason. It shouldn't matter. Chummer5isalive says it shouldn't matter. We're both confused as to why this matters. But since it aPpArEnTlY matters, I'm fixing it on this end this way so you can buy reagents in Point Buy again.